requirements.txt
cmake
djangorestframework
mongoengine==0.18.2
python-dateutil
Pyjwt
Django==2.2.4
django-cors-headers
nose
pymongo==3.9.0
Sphinx==1.5.5
sphinx-rtd-theme==0.2.4
openpyxl==2.5.6
django-crontab==0.7.1
Pillow==5.2.0
numpy==1.17.2
Click==7.0
dnspython==2.0.0
face-recognition==1.3.0